@@ -7,7 +7,7 @@ use Doctrine\ORM\Mapping as ORM;
 /**
  * Avis
  *
- * @ORM\Table(name="avis", indexes={@ORM\Index(name="cle", columns={"id_conducteur"})})
+ * @ORM\Table(name="avis")
  * @ORM\Entity
  */
 class Avis
@@ -24,19 +24,45 @@ class Avis
     /**
      * @var int
      *
+     * @ORM\Column(name="id_client", type="integer", nullable=false)
+     */
+    private $idClient;
+
+    /**
+     * @var int
+     *
      * @ORM\Column(name="note", type="integer", nullable=false)
      */
     private $note;
 
-    /**
-     * @var \Conducteur
-     *
-     * @ORM\ManyToOne(targetEntity="Conducteur")
-     * @ORM\JoinColumns({
-     *   @ORM\JoinColumn(name="id_conducteur", referencedColumnName="id_conducteur")
-     * })
-     */
-    private $idConducteur;
+    public function getIdAvis(): ?int
+    {
+        return $this->idAvis;
+    }
+
+    public function getIdClient(): ?int
+    {
+        return $this->idClient;
+    }
+
+    public function setIdClient(int $idClient): self
+    {
+        $this->idClient = $idClient;
+
+        return $this;
+    }
+
+    public function getNote(): ?int
+    {
+        return $this->note;
+    }
+
+    public function setNote(int $note): self
+    {
+        $this->note = $note;
+
+        return $this;
+    }
 
 
 }
