@@ -2,6 +2,7 @@
 
 namespace App\Entity;
 
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 
 /**
@@ -50,31 +51,114 @@ class Contrat
     private $statut;
 
     /**
+     * @var int|null
+     *
+     * @ORM\Column(name="id_admin", type="integer", nullable=true)
+     */
+    private $idAdmin;
+
+    /**
+     * @var int|null
+     *
+     * @ORM\Column(name="id_conducteur", type="integer", nullable=true)
+     */
+    private $idConducteur;
+
+    /**
      * @var string
      *
      * @ORM\Column(name="qr_code", type="string", length=100, nullable=false)
      */
     private $qrCode;
 
-    /**
-     * @var \Admin
-     *
-     * @ORM\ManyToOne(targetEntity="Admin")
-     * @ORM\JoinColumns({
-     *   @ORM\JoinColumn(name="id_admin", referencedColumnName="id_admin")
-     * })
-     */
-    private $idAdmin;
+    public function getIdContrat(): ?int
+    {
+        return $this->idContrat;
+    }
 
-    /**
-     * @var \Conducteur
-     *
-     * @ORM\ManyToOne(targetEntity="Conducteur")
-     * @ORM\JoinColumns({
-     *   @ORM\JoinColumn(name="id_conducteur", referencedColumnName="id_conducteur")
-     * })
-     */
-    private $idConducteur;
+    public function getDateDebut(): ?\DateTimeInterface
+    {
+        return $this->dateDebut;
+    }
+
+    public function setDateDebut(\DateTimeInterface $dateDebut): self
+    {
+        $this->dateDebut = $dateDebut;
+
+        return $this;
+    }
+
+    public function getDateFin(): ?\DateTimeInterface
+    {
+        return $this->dateFin;
+    }
+
+    public function setDateFin(\DateTimeInterface $dateFin): self
+    {
+        $this->dateFin = $dateFin;
+
+        return $this;
+    }
+
+    public function getPrix(): ?int
+    {
+        return $this->prix;
+    }
+
+    public function setPrix(int $prix): self
+    {
+        $this->prix = $prix;
+
+        return $this;
+    }
+
+    public function getStatut(): ?string
+    {
+        return $this->statut;
+    }
+
+    public function setStatut(string $statut): self
+    {
+        $this->statut = $statut;
+
+        return $this;
+    }
+
+    public function getIdAdmin(): ?int
+    {
+        return $this->idAdmin;
+    }
+
+    public function setIdAdmin(?int $idAdmin): self
+    {
+        $this->idAdmin = $idAdmin;
+
+        return $this;
+    }
+
+    public function getIdConducteur(): ?int
+    {
+        return $this->idConducteur;
+    }
+
+    public function setIdConducteur(?int $idConducteur): self
+    {
+        $this->idConducteur = $idConducteur;
+
+        return $this;
+    }
+
+    public function getQrCode(): ?string
+    {
+        return $this->qrCode;
+    }
+
+    public function setQrCode(string $qrCode): self
+    {
+        $this->qrCode = $qrCode;
+
+        return $this;
+    }
 
 
 }
