@@ -4,32 +4,44 @@ namespace App\Entity;
 
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
-use App\Repository\ReclamationRepository;
-use Symfony\Component\Validator\Constraints as Assert;
 
-
-#[ORM\Entity(repositoryClass: ReclamationRepository::class)]
+/**
+ * Reclamation
+ *
+ * @ORM\Table(name="reclamation")
+ * @ORM\Entity
+ */
 class Reclamation
 {
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column]
-    private ?int $id = null;
+    /**
+     * @var int
+     *
+     * @ORM\Column(name="id", type="integer", nullable=false)
+     * @ORM\Id
+     * @ORM\GeneratedValue(strategy="IDENTITY")
+     */
+    private $id;
 
-    #[ORM\Column(length: 255)]
-    private ?string $type = null;
+    /**
+     * @var string
+     *
+     * @ORM\Column(name="type", type="string", length=255, nullable=false)
+     */
+    private $type;
 
-    #[ORM\Column(type: Types::DATE_MUTABLE)]
-    private ?\DateTimeInterface $dateR = null;
+    /**
+     * @var \DateTime
+     *
+     * @ORM\Column(name="date_r", type="date", nullable=false)
+     */
+    private $dateR;
 
-    #[ORM\Column(length: 255)]
-    #[Assert\NotBlank(message:"Description reclamation doit etre non vide")]
-     #[Assert\Length(min:7,max:10000, minMessage:"Doit etre > 7.", maxMessage:"Doit etre <=10000")]
-    private ?string $description = null;
-
-    
-    #[ORM\ManyToOne(inversedBy: 'reclamations')]
-    private ?Client $idClient = null;
+    /**
+     * @var string
+     *
+     * @ORM\Column(name="description", type="string", length=255, nullable=false)
+     */
+    private $description;
 
     public function getId(): ?int
     {
@@ -68,18 +80,6 @@ class Reclamation
     public function setDescription(string $description): self
     {
         $this->description = $description;
-
-        return $this;
-    }
-
-    public function getIdClient(): ?Client
-    {
-        return $this->idClient;
-    }
-
-    public function setIdClient(?Client $idClient): self
-    {
-        $this->idClient = $idClient;
 
         return $this;
     }
